requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0

# HuggingFace
transformers>=4.35.0

# Medical Imaging
pydicom>=2.4.0

# COCO tools
pycocotools>=2.0.6

# Computer Vision
supervision==0.3.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Benchmarking / ONNX
onnx>=1.14.0
onnxruntime>=1.16.0          # CPU inference
# onnxruntime-gpu>=1.16.0    # Uncomment for GPU inference

# Utilities
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
pandas>=2.0.0
timm>=0.9.0
tqdm>=4.65.0

# Notebook
jupyter>=1.0.0
ipywidgets>=8.0.0
